Fix for allowing deletion of tags in DRUF
Patching `package_patch` ssuch that if a tag_string is part of
the patch, it removes existing tags from the dict so that the
tags will be recreated based on the tag_string.

Without doing this, when a user is trying to remove
tags they don't get deleted.

Need to investigate whether there is a better place to do this and/or
whether and where this should be submitted as a PR.